admin-view-micro-1/requirements.txt
Flask==2.3.2
boto3==1.26.153
python-dotenv==1.0.0
prometheus_client
werkzeug
mysql-connector-python